+++ unikmer/cmd/concat.go
@@ -60,7 +60,7 @@ Attentions:
 			}
 		}
 
-		checkFileSuffix(extDataFile, files...)
+		checkFileSuffix(opt, extDataFile, files...)
 
 		outFile := getFlagString(cmd, "out-prefix")
 		sortedKmers := getFlagBool(cmd, "sorted")