+++ unikmer/cmd/rfilter.go
@@ -75,7 +75,7 @@ Rank file:
 			}
 		}
 
-		checkFileSuffix(extDataFile, files...)
+		checkFileSuffix(opt, extDataFile, files...)
 
 		outFile := getFlagString(cmd, "out-prefix")