+++ unikmer/cmd/sample.go
@@ -62,7 +62,7 @@ Attentions:
 			}
 		}
 
-		checkFileSuffix(extDataFile, files...)
+		checkFileSuffix(opt, extDataFile, files...)
 
 		outFile := getFlagString(cmd, "out-prefix")