+++ unikmer/cmd/split.go
@@ -88,7 +88,7 @@ Tips:
 			}
 		}
 
-		checkFileSuffix(extDataFile, files...)
+		checkFileSuffix(opt, extDataFile, files...)
 
 		var m []uint64
 		var taxondb *unikmer.Taxonomy